# api-gateway/api/docs/swagger.yaml
definitions:
...
  models.Address:
    properties:
      country:
        type: string
      id:
        type: string
      owner_id:
        type: string
      street:
        type: string
    type: object
...
  models.Error:
    properties:
      code:
        type: integer
      description:
        type: string
      error: {}
    type: object
...
  models.ReviewUpdate:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      rating:
        type: integer
    type: object
  models.VerifyResponse:
    properties:
      accsee_token:
        type: string
      addresses:
        items:
          $ref: '#/definitions/models.Address'
        type: array
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      password:
        type: string
      refresh_token:
        type: string
    type: object
...
  review.Empty:
    type: object
  review.ReviewResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      post_id:
        type: string
      rating:
        type: integer
      udpated_at:
        type: string
    type: object
  review.Reviews:
    properties:
      reviews:
        items:
          $ref: '#/definitions/review.ReviewResponse'
        type: array
    type: object
host: localhost:9090
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is exam server api server
  termsOfService: 2 term exam
  title: exam api
  version: "2.0"
paths:
...
  /v1/review/delete/{id}:
    delete:
      consumes:
      - application/json
      description: this deleting information of review
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.Empty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete Review
      tags:
      - Review
  /v1/review/get/{id}:
    get:
      consumes:
      - application/json
      description: this will display the review information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.Reviews'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get Review
      tags:
      - Review
  /v1/review/update:
    put:
      consumes:
      - application/json
      description: this updating information of review
      parameters:
      - description: Review
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ReviewUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.ReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update Review
      tags:
      - Review
  /v1/verify/{email}/{code}:
    get:
      consumes:
      - application/json
      description: Verifys customer
      parameters:
      - description: email
        in: path
        name: email
        required: true
        type: string
      - description: code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VerifyResponse'
      summary: Verify customer
      tags:
      - Customer
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"